+++ Runtime/Core/HerdKVFlushMode.cs
@@ -1,0 +1,9 @@
+namespace Degubites.HerdKV
+{
+public enum HerdKVFlushMode
+{
+    Manual = 0,
+    FlushOnWrite = 1,
+    WriteThrough = 2
+}
+}